backend/requirements.txt
# Runtime dependencies for the PatentPath FastAPI backend.
alembic==1.13.3
anthropic==0.46.0
asyncpg==0.30.0
celery==5.3.6
dogpile.cache==1.3.3
fastapi==0.115.0
httpx==0.27.2
jinja2==3.1.4
numpy==2.1.2
passlib[bcrypt]==1.7.4
pydantic==2.9.2
pydantic-settings==2.6.0
python-jose[cryptography]==3.3.0
python-multipart==0.0.12
redis==5.1.1
scikit-learn==1.5.2
sentence-transformers==3.2.1
spacy==3.8.2
sqlalchemy==2.0.36
uvicorn[standard]==0.31.0
weasyprint==62.3
rank-bm25==0.2.2
python-epo-ops-client==4.2.1
